Move baseListPropsToMapJoinArgs
